feat!: follow conventional commits and setup automated releases (#48)
Closes #38 and #43.

With bors configured in this repository the flow is a bit different, but I think
this should be a good setup to allow automating releases. Instead of requiring
commit messages to follow the Conventional Commits specification, we just
require the PR title to follow it, since that is what ends up being the commit
message of the squash merge commit.

Overlaps with #47 so closing that one.